+++ ts/packages/spl-memo/src/coder/index.ts
@@ -3,7 +3,6 @@ import { Idl, Coder } from "@coral-xyz/anchor";
 import { SplMemoAccountsCoder } from "./accounts";
 import { SplMemoEventsCoder } from "./events";
 import { SplMemoInstructionCoder } from "./instructions";
-import { SplMemoStateCoder } from "./state";
 import { SplMemoTypesCoder } from "./types";
 
 /**
@@ -13,14 +12,12 @@ export class SplMemoCoder implements Coder {
   readonly accounts: SplMemoAccountsCoder;
   readonly events: SplMemoEventsCoder;
   readonly instruction: SplMemoInstructionCoder;
-  readonly state: SplMemoStateCoder;
   readonly types: SplMemoTypesCoder;
 
   constructor(idl: Idl) {
     this.accounts = new SplMemoAccountsCoder(idl);
     this.events = new SplMemoEventsCoder(idl);
     this.instruction = new SplMemoInstructionCoder(idl);
-    this.state = new SplMemoStateCoder(idl);
     this.types = new SplMemoTypesCoder(idl);
   }
 }